Mitigate peer-tag bug
This is still just part of a larger act of fixing the peer_tag crash, but when it does happen to me, ;afk is usually the culprit, since it checks the room id in any room it enters.

Here it'll only check the roomid if it's already in a room with the same title as 9610.